requirements.txt
# Data
pandas>=2.0.0
numpy>=1.26.0
pyarrow>=15.0.0

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0

# ML
scikit-learn>=1.4.0
lightgbm>=4.3.0
shap>=0.44.0

# Experiment tracking
mlflow>=2.12.0

# API
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# Monitoring
evidently>=0.4.0

# Utils
python-dotenv>=1.0.0
pydantic>=2.6.0
loguru>=0.7.0
tqdm>=4.66.0

# Dev
pytest>=8.0.0
jupyter>=1.0.0